testing/web-platform/tests/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>the blockquote element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/grouping-content.html#the-blockquote-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <h1>Description</h1>
    <p>This test validates the blockquote element.</p>

    <div id="log"></div>

    <hr>

    <blockquote>0- Cite is not provided.</blockquote>
    <blockquote cite="blehblah">1- Cite = "blehblah" </blockquote>
    <blockquote cite="http://{{host}}/">2- Cite = "http://{{host}}/"</blockquote>
    <blockquote cite="http://{{domains[www2]}}/">3- Cite = "http://{{domains[www2]}}/"</blockquote>
    <blockquote cite="http://天気の良い日.{{host}}/">4- Cite = "http://天気の良い日.{{host}}/"</blockquote>
    <blockquote cite="http://élève.{{host}}/">5- Cite = "http://élève.{{host}}/"</blockquote>
    <blockquote cite="https://{{domains[www]}}/">6- Cite = "https://{{domains[www]}}/"</blockquote>
    <blockquote cite=" http://{{host}}/ ">7- Cite = "http://{{host}}/" w/ spaces </blockquote>
    <blockquote cite="	http://{{host}}/	">8- Cite = "http://{{host}}/" w/ tabs </blockquote>
    <blockquote cite="&#x20http://{{host}}/&#x20">9- Cite = "http://{{host}}/" w/ Unicode space </blockquote>
    <blockquote cite="&#x09http://{{host}}/&#x09">10- Cite = "http://{{host}}/" w/ Unicode tab</blockquote>
    <blockquote cite="&#x0Ahttp://{{host}}/&#x0A">11- Cite = "http://{{host}}/" w/ Unicode newline</blockquote>
    <blockquote cite="&#x0Chttp://{{host}}/&#x0C">12- Cite = "http://{{host}}/" w/ Unicode formfeed</blockquote>
    <blockquote cite="&#x0Dhttp://{{host}}/&#x0D">13- Cite = "http://{{host}}/" w/ Unicode carriage return </blockquote>
    <blockquote cite="  &#x20&#x09&#x0A http://{{host}}/  &#x0C 	&#x0D   ">14- Cite = "http://{{host}}/" w/ combo spaces </blockquote>

    <script>
        "use strict";

        var testURLs = [{actual: "", resolved: ""},
                {actual: "blehblah", resolved: document.location.protocol + "//" + document.location.host + "/blehblah"},
                {actual: "http://{{host}}/", resolved: "http://{{host}}/"},
                {actual: "http://{{domains[www2]}}/", resolved: "http://{{domains[www2]}}/"},
                {actual: "http://天気の良い日.{{host}}/", resolved: "http://天気の良い日.{{host}}/"},
                {actual: "http://élève.{{host}}/", resolved: "http://élève.{{host}}/"},
                {actual: "https://{{domains[www]}}/", resolved: "https://{{domains[www]}}/"},
                {actual: " http://{{host}}/ ", resolved: "http://{{host}}/"}, // with spaces
                {actual: "	http://{{host}}/	", resolved: "http://{{host}}/"}, // with tabs
                {actual: "\u0020http://{{host}}/\u0020", resolved: "http://{{host}}/"}, // with unicode space
                {actual: "\u0009http://{{host}}/\u0009", resolved: "http://{{host}}/"}, // with unicode tab
                {actual: "\u000Ahttp://{{host}}/\u000A", resolved: "http://{{host}}/"}, // with unicode newline
                {actual: "\u000Chttp://{{host}}/\u000C", resolved: "http://{{host}}/"}, // with unicode formfeed
                {actual: "\u000Dhttp://{{host}}/\u000D", resolved: "http://{{host}}/"}, // with unicode carriage return
                {actual: "  \u0020\u0009\u000A http://{{host}}/  \u000C 	\u000D   ", resolved: "http://{{host}}/"} // with combo unicode spaces
                ];

        var testElement = {}, testQuotes = [];

        setup(function () {
            testQuotes = document.getElementsByTagName("blockquote");
            testElement = testQuotes[0];
        });

        // check that prototype matches spec's DOM interface
        test(function () {
            assert_equals(Object.getPrototypeOf(testElement), HTMLQuoteElement.prototype, "HTMLQuoteElement.prototype should be used for blockquote.");
        }, "The prototype for blockquote is HTMLQuoteElement.prototype");

        // check that "cite" is an "own" property
        test(function () {
            assert_own_property(testElement, "cite", "blockquote should have 'cite' as own property.");
        }, "blockquote should have 'cite' as own property.");

        // If the content attribute is absent, the IDL attribute must return the default value, if the content attribute has one, or else the empty string.
        test(function () {
            assert_equals(testElement.cite, "", "If the cite content attribute is absent, the IDL attribute must return the empty string.");
        }, "If the cite content attribute is absent, the IDL attribute must return the empty string.");

        // On setting, the IDL attribute must set the content attribute to the specified literal value.
        test(function () {
            testElement.cite = "foo bar";
            assert_equals(testElement.attributes["cite"].value, "foo bar", "Setting cite IDL sets content attribute to specified literal value.");
        }, "Setting cite IDL sets content attribute to specified literal value.");

        // All other testQuote examples have content attributes to test vis-a-vis registering and resolving

        for (var i = 1; i < testQuotes.length; i++) {
            // Check that element's cite content attribute registered properly (literally, not resolved)
            test(function () {
                assert_equals(testQuotes[i].attributes["cite"].value, testURLs[i].actual, "blockquote #" + i + ": content attribute");
            }, "blockquote #" + i + ": cite content attribute registers markup.");
        }

        for (var i = 1; i < testQuotes.length; i++) {
            // The cite IDL attribute must reflect the element's cite content attribute (resolving URLs relative to the element)...
            test(function () {
                assert_equals(testQuotes[i].cite, testURLs[i].resolved, "blockquote #" + i + ": cite IDL reflects content attribute.");
            }, "blockquote #" + i + ": cite IDL reflects content attribute.");
        }

        // Not checking: Is a sectioning root

        </script>
</body>
</html>
