w3c-reports/12092014/mac-wk2/runned-tests/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub-expected.txt

Description

This test validates the blockquote element.

0- Cite is not provided.
1- Cite = "blehblah"
2- Cite = "http://localhost/"
3- Cite = "http://www2.localhost/"
4- Cite = "http://天気の良い日.localhost/"
5- Cite = "http://élève.localhost/"
6- Cite = "https://www.localhost/"
7- Cite = "http://localhost/" w/ spaces
8- Cite = "http://localhost/" w/ tabs
9- Cite = "http://localhost/" w/ Unicode space
10- Cite = "http://localhost/" w/ Unicode tab
11- Cite = "http://localhost/" w/ Unicode newline
12- Cite = "http://localhost/" w/ Unicode formfeed
13- Cite = "http://localhost/" w/ Unicode carriage return
14- Cite = "http://localhost/" w/ combo spaces

PASS The prototype for blockquote is HTMLQuoteElement.prototype 
FAIL blockquote should have 'cite' as own property. assert_own_property: blockquote should have 'cite' as own property. expected property "cite" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:69:32
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:68:13)
PASS If the cite content attribute is absent, the IDL attribute must return the empty string. 
PASS Setting cite IDL sets content attribute to specified literal value. 
PASS blockquote #1: cite content attribute registers markup. 
PASS blockquote #2: cite content attribute registers markup. 
PASS blockquote #3: cite content attribute registers markup. 
PASS blockquote #4: cite content attribute registers markup. 
PASS blockquote #5: cite content attribute registers markup. 
PASS blockquote #6: cite content attribute registers markup. 
PASS blockquote #7: cite content attribute registers markup. 
PASS blockquote #8: cite content attribute registers markup. 
PASS blockquote #9: cite content attribute registers markup. 
PASS blockquote #10: cite content attribute registers markup. 
PASS blockquote #11: cite content attribute registers markup. 
PASS blockquote #12: cite content attribute registers markup. 
PASS blockquote #13: cite content attribute registers markup. 
PASS blockquote #14: cite content attribute registers markup. 
FAIL blockquote #1: cite IDL reflects content attribute. assert_equals: blockquote #1: cite IDL reflects content attribute. expected "http://localhost:8800/blehblah" but got "http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/blehblah"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:95:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:94:17)
PASS blockquote #2: cite IDL reflects content attribute. 
PASS blockquote #3: cite IDL reflects content attribute. 
FAIL blockquote #4: cite IDL reflects content attribute. assert_equals: blockquote #4: cite IDL reflects content attribute. expected "http://天気の良い日.localhost/" but got "http://xn--n8j6ds53lwwkrqhv28a.localhost/"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:95:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:94:17)
FAIL blockquote #5: cite IDL reflects content attribute. assert_equals: blockquote #5: cite IDL reflects content attribute. expected "http://élève.localhost/" but got "http://xn--lve-6lad.localhost/"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:95:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-blockquote-element/grouping-blockquote.sub.html:94:17)
PASS blockquote #6: cite IDL reflects content attribute. 
PASS blockquote #7: cite IDL reflects content attribute. 
PASS blockquote #8: cite IDL reflects content attribute. 
PASS blockquote #9: cite IDL reflects content attribute. 
PASS blockquote #10: cite IDL reflects content attribute. 
PASS blockquote #11: cite IDL reflects content attribute. 
PASS blockquote #12: cite IDL reflects content attribute. 
PASS blockquote #13: cite IDL reflects content attribute. 
PASS blockquote #14: cite IDL reflects content attribute. 
